update tmux conf
having an issue with the original way of navigating between
tmux and vim splits. I was getting error messages when I couldn't
navigate a certain direction. Though the error was just the echoing
of the command that failed, and it cleared my whole terminal.
Not so great. But this fixes that.